backend/requirements.txt
fastapi==0.115.12
numpy==2.3.0
python-multipart==0.0.6
Pillow==11.2.1
tensorflow==2.19.0
uvicorn==0.34.3
gunicorn